消除几个警告

+++ HighlightedSearch/Classes/PinyinFormatter.m
@@ -79,19 +79,19 @@ static unichar numericValues[] = {
   NSString *lowerCasePinyinStr = [pinyinStr lowercaseString];
   if ([lowerCasePinyinStr matchesPatternRegexPattern:@"[a-z]*[1-5]?"]) {
     unichar defautlCharValue = '$';
-    int defautlIndexValue = -1;
+    NSInteger defautlIndexValue = -1;
     unichar unmarkedVowel = defautlCharValue;
-    int indexOfUnmarkedVowel = defautlIndexValue;
+    NSInteger indexOfUnmarkedVowel = defautlIndexValue;
     unichar charA = 'a';
     unichar charE = 'e';
     NSString *ouStr = @"ou";
     NSString *allUnmarkedVowelStr = @"aeiouv";
     NSString *allMarkedVowelStr = @"āáăàaēéĕèeīíĭìiōóŏòoūúŭùuǖǘǚǜü";
     if ([lowerCasePinyinStr matchesPatternRegexPattern:@"[a-z]*[1-5]"]) {
-        int tuneNumber = [PinyinFormatter getNumericValue:[lowerCasePinyinStr characterAtIndex:lowerCasePinyinStr.length -1]];
-      int indexOfA = [lowerCasePinyinStr indexOf:charA];
-      int indexOfE = [lowerCasePinyinStr indexOf:charE];
-      int ouIndex = [lowerCasePinyinStr indexOfString:ouStr];
+        NSInteger tuneNumber = [PinyinFormatter getNumericValue:[lowerCasePinyinStr characterAtIndex:lowerCasePinyinStr.length -1]];
+      NSInteger indexOfA = [lowerCasePinyinStr indexOf:charA];
+      NSInteger indexOfE = [lowerCasePinyinStr indexOf:charE];
+      NSInteger ouIndex = [lowerCasePinyinStr indexOfString:ouStr];
       if (-1 != indexOfA) {
         indexOfUnmarkedVowel = indexOfA;
         unmarkedVowel = charA;
@@ -105,7 +105,7 @@ static unichar numericValues[] = {
         unmarkedVowel = [ouStr characterAtIndex:0];
       }
       else {
-        for (int i = [lowerCasePinyinStr length] - 1; i >= 0; i--) {
+        for (NSInteger i = [lowerCasePinyinStr length] - 1; i >= 0; i--) {
           if ([[NSString valueOfChar:[lowerCasePinyinStr characterAtIndex:i]] matchesPatternRegexPattern:@"[aeiouv]"]) {
             indexOfUnmarkedVowel = i;
             unmarkedVowel = [lowerCasePinyinStr characterAtIndex:i];
@@ -114,9 +114,9 @@ static unichar numericValues[] = {
         }
       }
       if ((defautlCharValue != unmarkedVowel) && (defautlIndexValue != indexOfUnmarkedVowel)) {
-        int rowIndex = [allUnmarkedVowelStr indexOf:unmarkedVowel];
-        int columnIndex = tuneNumber - 1;
-        int vowelLocation = rowIndex * 5 + columnIndex;
+        NSInteger rowIndex = [allUnmarkedVowelStr indexOf:unmarkedVowel];
+        NSInteger columnIndex = tuneNumber - 1;
+        NSInteger vowelLocation = rowIndex * 5 + columnIndex;
         unichar markedVowel = [allMarkedVowelStr characterAtIndex:vowelLocation];
         NSMutableString *resultBuffer = [[NSMutableString alloc] init];
           [resultBuffer appendString:[[lowerCasePinyinStr substringToIndex:indexOfUnmarkedVowel+1] stringByReplacingOccurrencesOfString:@"v" withString:@"ü"]];
